refactor: SavedDataLoad의 패키지를 domain으로 옮겨라
- 데이터를 읽어와 도메인을 만드는 로직을 가지고 있기 때문에 충분히 도메인 로직이라 판단하였다.
